Add namespace to more math functions

+++ include/geometryCheck.h
@@ -150,34 +150,34 @@ struct geometry_check {
         float vy0 = 0.0;
         if (theta <= boundaryVector[nLines].y1) {
           particlesPointer->xprevious[indx] =
-              r * cos(boundaryVector[nLines].y2 + theta);
+              r * std::cos(boundaryVector[nLines].y2 + theta);
           particlesPointer->yprevious[indx] =
-              r * sin(boundaryVector[nLines].y2 + theta);
+              r * std::sin(boundaryVector[nLines].y2 + theta);
           particlesPointer->x[indx] =
-              rprev * cos(boundaryVector[nLines].y2 + theta);
+              rprev * std::cos(boundaryVector[nLines].y2 + theta);
           particlesPointer->y[indx] =
-              rprev * sin(boundaryVector[nLines].y2 + theta);
+              rprev * std::sin(boundaryVector[nLines].y2 + theta);
 
-          vx0 = vrMag * cos(boundaryVector[nLines].y2 + theta) -
-                vthetaMag * sin(boundaryVector[nLines].y2 + theta);
-          vy0 = vrMag * sin(boundaryVector[nLines].y2 + theta) +
-                vthetaMag * cos(boundaryVector[nLines].y2 + theta);
+          vx0 = vrMag * std::cos(boundaryVector[nLines].y2 + theta) -
+                vthetaMag * std::sin(boundaryVector[nLines].y2 + theta);
+          vy0 = vrMag * std::sin(boundaryVector[nLines].y2 + theta) +
+                vthetaMag * std::cos(boundaryVector[nLines].y2 + theta);
           particlesPointer->vx[indx] = vx0;
           particlesPointer->vy[indx] = vy0;
         } else if (theta >= boundaryVector[nLines].y2) {
           particlesPointer->xprevious[indx] =
-              rprev * cos(thetaPrev - boundaryVector[nLines].y2);
+              rprev * std::cos(thetaPrev - boundaryVector[nLines].y2);
           particlesPointer->yprevious[indx] =
-              rprev * sin(thetaPrev - boundaryVector[nLines].y2);
+              rprev * std::sin(thetaPrev - boundaryVector[nLines].y2);
           particlesPointer->x[indx] =
-              r * cos(theta - boundaryVector[nLines].y2);
+              r * std::cos(theta - boundaryVector[nLines].y2);
           particlesPointer->y[indx] =
-              r * sin(theta - boundaryVector[nLines].y2);
+              r * std::sin(theta - boundaryVector[nLines].y2);
 
-          vx0 = vrMag * cos(theta - boundaryVector[nLines].y2) -
-                vthetaMag * sin(theta - boundaryVector[nLines].y2);
-          vy0 = vrMag * sin(theta - boundaryVector[nLines].y2) +
-                vthetaMag * cos(theta - boundaryVector[nLines].y2);
+          vx0 = vrMag * std::cos(theta - boundaryVector[nLines].y2) -
+                vthetaMag * std::sin(theta - boundaryVector[nLines].y2);
+          vy0 = vrMag * std::sin(theta - boundaryVector[nLines].y2) +
+                vthetaMag * std::cos(theta - boundaryVector[nLines].y2);
           particlesPointer->vx[indx] = vx0;
           particlesPointer->vy[indx] = vy0;
         }
@@ -306,12 +306,12 @@ struct geometry_check {
       float dr = closeGeomGridr[rHashInd + 1] - closeGeomGridr[rHashInd];
       float dz = closeGeomGridz[zHashInd + 1] - closeGeomGridz[zHashInd];
       float dy = closeGeomGridy[yHashInd + 1] - closeGeomGridy[yHashInd];
-      int rInd = floor((r_position - closeGeomGridr[rHashInd]) / dr + 0.5f);
-      int zInd = floor(
+      int rInd = std::floor((r_position - closeGeomGridr[rHashInd]) / dr + 0.5f);
+      int zInd = std::floor(
           (particlesPointer->zprevious[indx] - closeGeomGridz[zHashInd]) / dz +
           0.5f);
       int i = 0;
-      int yInd = floor(
+      int yInd = std::floor(
           (particlesPointer->yprevious[indx] - closeGeomGridy[yHashInd]) / dy +
           0.5f);
       // std::cout << "rHashInd " << rHashInd << " " << yHashInd << " " <<
@@ -544,8 +544,8 @@ struct geometry_check {
 #endif
       float dr = closeGeomGridr[1] - closeGeomGridr[0];
       float dz = closeGeomGridz[1] - closeGeomGridz[0];
-      int rInd = floor((r_position - closeGeomGridr[0]) / dr + 0.5f);
-      int zInd = floor(
+      int rInd = std::floor((r_position - closeGeomGridr[0]) / dr + 0.5f);
+      int zInd = std::floor(
           (particlesPointer->zprevious[indx] - closeGeomGridz[0]) / dz + 0.5f);
       if (rInd < 0 || rInd >= nR_closeGeom[0])
         rInd = 0;
@@ -817,12 +817,12 @@ struct geometry_check {
           thetaNew = theta0 + (intersectionx[minDistInd] - pdim1previous) /
                                   (pdim1 - pdim1previous) * (theta1 - theta0);
           particlesPointer->yprevious[indx] =
-              intersectionx[minDistInd] * sinf(thetaNew);
+              intersectionx[minDistInd] * std::sin(thetaNew);
           particlesPointer->y[indx] = particlesPointer->yprevious[indx];
           // particlesPointer->y[indx] =
           // intersectionx[minDistInd]*cosf(thetaNew);
           particlesPointer->x[indx] =
-              intersectionx[minDistInd] * cosf(thetaNew);
+              intersectionx[minDistInd] * std::cos(thetaNew);
 #else
           // std::cout << "Particle index " << indx << " hit wall and is
           // calculating y point " << particlesPointer->yprevious[indx] << " " <<
@@ -889,13 +889,13 @@ struct geometry_check {
 
         partDotNormal =
             vectorDotProduct(particleTrackVector, surfaceNormalVector);
-        thetaImpact = acos(partDotNormal);
+        thetaImpact = std::acos(partDotNormal);
         if (thetaImpact > 3.14159265359 * 0.5) {
           thetaImpact = std::abs(thetaImpact - (3.14159265359));
         }
         thetaImpact = thetaImpact * 180.0 / 3.14159265359;
-        EdistInd = floor((E0 - E0dist) / dEdist);
-        AdistInd = floor((thetaImpact - A0dist) / dAdist);
+        EdistInd = std::floor((E0 - E0dist) / dEdist);
+        AdistInd = std::floor((thetaImpact - A0dist) / dAdist);
         int surfaceHit =
             boundaryVector[particlesPointer->wallHit[indx]].surfaceNumber;
         int surface = boundaryVector[particlesPointer->wallHit[indx]].surface;
